third_party/blink/web_tests/svg/as-image/svg-image-usecounters-in-image.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
// From third_party/blink/public/mojom/use_counter/metrics/web_feature.mojom
var SVGSMILAnimationInImageRegardlessOfCache = 768;
var SVGImage = 5492;

window.jsTestIsAsync = true;

shouldBeFalse("internals.isUseCounted(document, SVGSMILAnimationInImageRegardlessOfCache)");
shouldBeFalse("internals.isUseCounted(document, SVGImage)");

let static_image = new Image();
static_image.onload = function() {
    shouldBeFalse("internals.isUseCounted(document, SVGSMILAnimationInImageRegardlessOfCache)");
    shouldBeTrue("internals.isUseCounted(document, SVGImage)");

    let animated_image = new Image();
    animated_image.onload = function() {
        shouldBeTrue("internals.isUseCounted(document, SVGSMILAnimationInImageRegardlessOfCache)");
        shouldBeTrue("internals.isUseCounted(document, SVGImage)");
        finishJSTest();
    }
    animated_image.src = "resources/animated-rect-color.svg";
}
static_image.src = "resources/100px-green-rect.svg";
</script>
</html>
